Move grid utilities to their own module
